+++ app/services/forest_liana/has_many_dissociator.rb
@@ -14,17 +14,22 @@ module ForestLiana
 
       remove_association = !@with_deletion || @association.macro == :has_and_belongs_to_many
 
-      if remove_association
-        if @data.is_a?(Array)
-          @data.each do |record_deleted|
-            associated_records.delete(@association.klass.find(record_deleted[:id]))
-          end
-        end
+      if @data.is_a?(Array)
+        record_ids = @data.map { |record| record[:id] }
+      elsif @data.dig('attributes').present?
+        record_ids = ForestLiana::ResourcesGetter.get_ids_from_request(@params)
+      else
+        record_ids = Array.new
       end
 
-      if @with_deletion
-        if @data.is_a?(Array)
-          record_ids = @data.map { |record| record[:id] }
+      if !record_ids.nil? && record_ids.any?
+        if remove_association
+          record_ids.each do |id|
+            associated_records.delete(@association.klass.find(id))
+          end
+        end
+
+        if @with_deletion
           record_ids = record_ids.select { |record_id| @association.klass.exists?(record_id) }
           @association.klass.destroy(record_ids)
         end