+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
 ## [Unreleased]
+### Added
+- Resources getter - Add a get_ids_from_request method to get all models IDs given a query or an ID list.
+- Resource Deletion - Users can now bulk delete records.
+
 ### Fixed
 - Has Many Relationships - Fix records count, i.e consider filters when counting.